CORE: Update IbgeCode type from long to int

+++ ChallengeIBGE.Core/Contexts/AddressContext/Entities/Address.cs
@@ -4,7 +4,7 @@ namespace ChallengeIBGE.Core.Contexts.AddressContext.Entities;
 
 public class Address : Entity
 {
-    public Address(string city, string state, long ibgeCode)
+    public Address(string city, string state, int ibgeCode)
     {
         City = city;
         State = state;
@@ -12,9 +12,9 @@ public class Address : Entity
     }
     public string City { get; private set; }
     public string State { get; private set; }
-    public long IbgeCode { get; private set; }
+    public int IbgeCode { get; private set; }
 
     public void UpdateCity(string city) => City = city;
     public void UpdateState(string state) => State = state;
-    public void UpdateIbgeCode(long ibgeCode) => IbgeCode = ibgeCode;
+    public void UpdateIbgeCode(int ibgeCode) => IbgeCode = ibgeCode;
 }
